Use ndk version of libpng & libjpeg on ImageMagick
ImageMagick is a set of static libraries built with NDK, and they use
libpng & libjpeg which are built without NDK. This potentially causes a
link-type error because NDK libraries should not link against platform
libraries. Static versions of libpng & libjpeg are built with NDK, so
linking against them resolves the problem.

Bug: 77560753
Test: m -j succeeded
Change-Id: Id69008e7fc4cee51f667cd344639d089fdf01315